+++ regression_model/config/config.py
@@ -7,7 +7,7 @@ import pandas as pd
 pd.options.display.max_rows = 10
 pd.options.display.max_columns = 10
 
-PACKAGE_ROOT = pathlib.Path(__file__).resolve().parent
+PACKAGE_ROOT = pathlib.Path(__file__).resolve().parent.parent
 TRAINED_MODEL_DIR = PACKAGE_ROOT / 'trained_models'
 DATASET_DIR = PACKAGE_ROOT / 'datasets'
