+++ regression_model/processing/features.py
@@ -1,7 +1,7 @@
 import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
 
-from errors import InvalidModelInputError
+from processing.errors import InvalidModelInputError
 
 
 class LogTransformer(BaseEstimator, TransformerMixin):
